+++ app/Http/Controllers/LogbookController.php
@@ -16,6 +16,8 @@ class LogbookController extends Controller
     public function index()
     {
 
+        $this->authorize('!koordinator');
+
         // mentor
         if (auth()->user()->level_id === 2) {
             $logbooks = DB::table('logbooks')
@@ -31,16 +33,33 @@ class LogbookController extends Controller
                 ->latest();
 
             return view('logbook.index', [
-                'logbooks' => $logbooks->paginate(5)->withQueryString(),
+                'logbooks' => $logbooks->paginate(5)
+            ]);
+        }
+
+        if (auth()->user()->level_id === 4) {
+            $logbooks = DB::table('logbooks')
+                ->join('judulprojeks', 'judulprojeks.id', 'logbooks.judul_id')
+                ->join('users', 'users.id', 'logbooks.user_id')
+                ->select('logbooks.*', 'judulprojeks.judul', 'judulprojeks.pembimbing', 'users.nama')
+                ->where('judul', 'like', '%' . request('search') . '%')
+                ->orWhere('nama', 'like', '%' . request('search') . '%')
+                ->orWhere('description', 'like', '%' . request('search') . '%')
+                ->orWhere('logbooks.created_at', 'like', '%' . request('search') . '%')
+                ->orWhere('logbooks.status', 'like', '%' . request('search') . '%')
+                ->latest();
+
+            return view('logbook.index', [
+                'logbooks' => $logbooks->paginate(5)
             ]);
         }
 
         // users / mahasiswa
-        $logbooks = Logbook::latest()->filter(request(['search']))->where('user_id', auth()->user()->id)->paginate(5)->withQueryString();
+        $logbooks = Logbook::with('judul')->latest()->filter(request(['search']))->where('user_id', auth()->user()->id)->paginate(5)->withQueryString();
 
         return view('logbook.index', [
             'logbooks' => $logbooks,
-            'status' => Logbook::where('user_id', auth()->user()->id)->where('status', 'diterima')->get(),
+            'status' => $logbooks->where('status', 'diterima')
         ]);
     }
 
@@ -49,6 +68,8 @@ class LogbookController extends Controller
      */
     public function create()
     {
+
+        $this->authorize('mahasiswa');
 
         return view('logbook.create', [
             'juduls' => Judulprojek::where('user_id', auth()->user()->id)->get()
@@ -60,6 +81,8 @@ class LogbookController extends Controller
      */
     public function store(Request $request)
     {
+        $this->authorize('mahasiswa');
+
         $validateData = $request->validate([
             'judul_id' => 'required',
             'description' => 'required',
@@ -95,6 +118,8 @@ class LogbookController extends Controller
      */
     public function edit($id)
     {
+        $this->authorize('pembimbing');
+
         return view('logbook.edit', [
             'logbook' =>  Logbook::find($id)
         ]);
@@ -105,6 +130,8 @@ class LogbookController extends Controller
      */
     public function update(Request $request, $id)
     {
+        $this->authorize('pembimbing');
+
         $validateData = $request->validate([
             'judul_id' => 'required',
             'description' => 'required',
@@ -123,6 +150,8 @@ class LogbookController extends Controller
      */
     public function destroy($id)
     {
+        $this->authorize('pembimbing');
+
         Logbook::destroy($id);
 
         Alert::success('Success!', 'Log Book Berhasil Dihapus');
